Handle PyTypes Config in own module
